feat(routes): - add crud routes for commands and categories - add login route for admin - add authorization middleware for admin - add controllers to handle all routes

+++ server/controllers/categoryController.js
@@ -1,24 +1,70 @@
-import models from '../models';
-
-const Category = models.Category;
-const Command = models.Command;
-
 class CategoryController {
-  static async getCategories(req, res) {
-    const categories = await Category.find({}).populate('commands');
-    res.json({ categories });
+  static saveCategory(models) {
+    const { Category } = models;
+    return async (req, res) => {
+      const { category } = req.body;
+      const createdCategory = await Category.create(category);
+      res.json({ success: true, category: createdCategory });
+    };
   }
 
-  static async deleteCategory(req, res) {
-    await Command.deleteMany({ category: req.params.id });
-    await Category.findOneAndDelete({ _id: req.params.id });
-    res.json({ success: true });
+  static getCategories(models) {
+    const { Category } = models;
+    return async (req, res) => {
+      const categories = await Category.find({}).populate('commands');
+      res.json({ success: true, categories });
+    };
   }
 
-  static async updateCategory(req, res) {
-    const { category, id } = req.params;
-    const updatedCategory = await Category.findOneAndUpdate({ _id: id }, category);
-    res.json({ category: updatedCategory });
+  static deleteCategory(models) {
+    const { Category, Command } = models;
+    return async (req, res) => {
+      const category = await Category.findOne({ _id: req.params.id });
+      if (category) {
+        await Command.deleteMany({ categoryId: req.params.id });
+        await Category.findOneAndDelete({ _id: req.params.id });
+        return res.json({ success: true });
+      }
+      return res.status(404).json({ success: false, error: 'category not found' });
+    };
+  }
+
+  static updateCategory(models) {
+    const { Category } = models;
+    return async (req, res) => {
+      const { id } = req.params;
+      const { category } = req.body;
+      const update = await Category.findOneAndUpdate({ _id: id }, category);
+      if (update) {
+        const updatedCategory = await Category.findOne({ _id: id });
+        return res.json({ success: true, category: updatedCategory });
+      }
+      return res.status(404).json({ success: false, error: 'category not found' });
+    };
+  }
+
+  static confirmCategory(models) {
+    const { Category } = models;
+    return async (req, res, next) => {
+      try {
+        const { categoryId } = req.body;
+        const category = await Category.findOne({ _id: categoryId });
+        if (category) return next();
+        return res.status(404).json({ success: false, error: 'category not found' });
+      } catch (e) {
+        return res.status(404).json({ success: false, error: 'category not found' });
+      }
+    };
+  }
+
+  static controllers(models) {
+    return {
+      getCategories: CategoryController.getCategories(models),
+      deleteCategory: CategoryController.deleteCategory(models),
+      updateCategory: CategoryController.updateCategory(models),
+      confirmCategory: CategoryController.confirmCategory(models),
+      saveCategory: CategoryController.saveCategory(models),
+    };
   }
 }
